requirements.txt
flask
python-dotenv
flask_sqlalchemy
psycopg2
werkzeug